+++ modules/react-native-status/android/src/main/java/im/status/ethereum/module/StatusPackage.java
@@ -30,6 +30,13 @@ public class StatusPackage implements ReactPackage {
         List<NativeModule> modules = new ArrayList<>();
 
         modules.add(new StatusModule(reactContext, this.rootedDevice));
+        modules.add(new AccountManager(reactContext));
+        modules.add(new EncryptionUtils(reactContext));
+        modules.add(new DatabaseManager(reactContext));
+        modules.add(new UIHelper(reactContext));
+        modules.add(new LogManager(reactContext));
+        modules.add(new Utils(reactContext));
+        modules.add(new NetworkManager(reactContext));
         modules.add(new RNSelectableTextInputModule(reactContext));
 
         return modules;
