+++ modules/react-native-status/ios/RCTStatus/NetworkManager.h
@@ -1,0 +1,9 @@
+#import <sys/utsname.h>
+#import <Foundation/Foundation.h>
+#import <React/RCTBridgeModule.h>
+#import "Statusgo.h"
+#import "RCTLog.h"
+
+@interface NetworkManager : NSObject <RCTBridgeModule>
+
+@end
