+++ modules/react-native-status/ios/RCTStatus/UIHelper.h
@@ -1,0 +1,8 @@
+#import <sys/utsname.h>
+#import <Foundation/Foundation.h>
+#import <React/RCTBridgeModule.h>
+#import "RCTLog.h"
+
+@interface UIHelper : NSObject <RCTBridgeModule>
+
+@end
